# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin、Gorm 和 Quasar 的低代码管理底座",
        "title": "Sweet Admin",
        "contact": {
            "name": "南",
            "email": "maxdwy@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/sweet_admin",
    "paths": {
...
        "/admin/file/download-url/{uuid}": {
            "get": {
                "description": "为当前用户可访问的文件生成短期签名下载地址",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "获取文件签名下载地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "有效期秒数，最大3600秒",
                        "name": "ttl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/download/{uuid}": {
            "get": {
                "description": "通过文件UUID下载文件（支持本地和OSS存储）",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/admin/file/preview-url/{uuid}": {
            "get": {
                "description": "为当前用户可访问的文件生成短期签名预览地址，用于富文本预览等无法携带 Authorization 头的场景",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "获取文件签名预览地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "有效期秒数，最大3600秒",
                        "name": "ttl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/preview/{uuid}": {
            "get": {
                "description": "通过文件UUID预览文件（图片、PDF等），支持本地和OSS存储",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "预览文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/admin/file/upload": {
            "post": {
                "description": "上传文件，支持秒传（MD5去重）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/upload/chunk": {
            "post": {
                "description": "上传单个文件分片，支持断点续传（幂等操作）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "上传分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片索引（从0开始）",
                        "name": "chunk_index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "分片文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/upload/init": {
            "post": {
                "description": "初始化分片上传，返回uploadId和分片信息。支持秒传检测。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "初始化分片上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "初始化参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ChunkUploadInitReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/upload/merge/{upload_id}": {
            "post": {
                "description": "所有分片上传完成后，调用此接口合并分片生成最终文件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "合并分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/upload/progress/{upload_id}": {
            "get": {
                "description": "获取分片上传进度，返回已上传的分片索引列表（用于断点续传）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "获取上传进度",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/file/{id}": {
            "get": {
                "description": "根据ID获取文件详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "获取文件信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除文件（同时删除存储中的物理文件）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/files/access/download/{uuid}": {
            "get": {
                "description": "通过短期签名 token 下载文件",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "签名下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "签名 token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.ChunkUploadInitReq": {
            "type": "object",
            "required": [
                "file_name",
                "file_size"
            ],
            "properties": {
                "file_md5": {
                    "description": "完整文件MD5（用于秒传检测）",
                    "type": "string"
                },
                "file_name": {
                    "description": "文件名",
                    "type": "string"
                },
                "file_size": {
                    "description": "文件总大小（字节）",
                    "type": "integer"
                },
                "file_type": {
                    "description": "文件MIME类型",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}